### 2025-02-21 #7832 @editorialbot
[PRE REVIEW]: mcsm-benchs: Benchmarking methods for multi-component signal processing
**Submitting author:** @jmiramont (<a href="http://orcid.org/0000-0002-3847-7811">J. M. Miramont</a>)
**Repository:** https://github.com/jmiramont/mcsm-benchs
**Branch with paper.md** (empty if default branch):
**Version:**  v0.1.2
