+++ paper.md
@@ -37,9 +37,9 @@ bibliography: paper.bib
 Time-frequency (TF) representations are natural encodings of non-stationary time series, also termed signals, helping to discern patterns that reveal their time-varying frequency structure [@flandrin1998time].
 The model one usually has in mind, when discussing TF representations is a so-called multi-component signal (MCS), where the latter is thought to be the sum of several components of individual interest.
 
-MCS processing is an area with a long history which, however, remains a field of methodological innovation [@bardenet2018zeros], [@colominas2020fully], [@kreme2020filtering], [@bardenet2021time], [@legros2022instantaneous], [@legros2022time],[@ghosh2022signal], [@pascal2022covariant], [@pascalfamille].
+MCS processing is an area with a long history which, however, remains a field of methodological innovation [@bardenet2018zeros; @colominas2020fully; @kreme2020filtering; @bardenet2021time; @legros2022instantaneous; @legros2022time; @ghosh2022signal; @pascal2022covariant; @pascalfamille].
 In such a context, systematically comparing existing methods while keeping a record of how they performed on a predefined set of representative problems, i.e. benchmarking, can shed light on novel avenues of research and set clear baselines for new approaches.
-This is a widely adopted strategy in neighboring fields such as optimization [@bartz2020benchmarking] [@hansen2021coco], [@benchopt] and machine learning [@mattson2020mlperf], yet there are no benchmarking tools for MCS processing.
+This is a widely adopted strategy in neighboring fields such as optimization [@bartz2020benchmarking; @hansen2021coco; @benchopt] and machine learning [@mattson2020mlperf], yet there are no benchmarking tools for MCS processing.
 `mcsm-benchs` is an open-source Python library for creating reproducible benchmarks of methods that aim at extracting information from MCSs.
 `Benchmark` objects can be created by simply passing a series of simulation parameters, a dictionary of methods and a set of performance metrics, while a `SignalBank` can be used to standardize comparisons across benchmarks.
 This latter can synthesize several MCSs as objects from a custom `Signal` class that behave like regular arrays in Python but also contain information about the components.
@@ -52,34 +52,11 @@ Several examples are given in the documentation, as well as a GitHub [repository
 
 `mcsm-benchs` brings to the table a common framework to easily carry out extensive comparisons between MCS-based approaches in a unified and objective way.
 For instance, `mcsm-benchs` was used to compare statistical methods for signal detection [@miramont2022public] and signal denoising of synthetic and real signals under different scenarios, such as white noise and real-world noises [@miramont2024].
-Additionally, benchmarks of instantaneous frequency and component estimation methods where introduced in [@miramont2023eusipco].
-<!-- What about public benchmarks? -->
+Additionally, benchmarks of instantaneous frequency and component estimation methods where introduced in @miramont2023eusipco.
+
 While `mcsm-benchs` was mainly designed with MCSs in mind, methods are always treated like black boxes, i.e. only the shape of the input and output arguments are constrained by the toolbox.
 Therefore, `mcsm-benchs` can potentially be used to compare any signal processing algorithm as long as its output matches the inputs of a given performance metric.
 Finally, to ease the adoption of our tool by the community, `mcsm-benchs` also supports methods coded in `Octave`/`Matlab`, so that these can be seamlessly integrated into new Python-based benchmarks.
-
-<!-- # Citations
-
-Citations to entries in paper.bib should be in
-[rMarkdown](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
-format.
-
[…]
