+++ paper.bib
@@ -4,7 +4,8 @@
   journal={Signal Processing},
   pages={109250},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.sigpro.2023.109250},
 }
 
 @inproceedings{pascalfamille,
@@ -21,7 +22,9 @@
   booktitle={ICASSP 2020-2020 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
   pages={5790--5794},
   year={2020},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.1109/icassp40776.2020.9053482},
+
 }
 
 @article{pascal2022covariant,
@@ -31,7 +34,8 @@
   pages={2950--2961},
   volume = {70},
   year={2022},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/tsp.2022.3181342},
 }
 
 @article{ghosh2022signal,
@@ -41,7 +45,8 @@
   volume={70},
   pages={1104--1117},
   year={2022},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/tsp.2022.3153596},
 }
 @article{laurent2021novel,
   title={A novel ridge detector for nonstationary multicomponent signals: Development and application to robust mode retrieval},
@@ -50,7 +55,8 @@
   volume={69},
   pages={3325--3336},
   year={2021},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/tsp.2021.3085113 },
 }
 
 @article{harmouche2017sliding,
@@ -61,13 +67,18 @@
   number={1},
   pages={251--263},
   year={2017},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/tsp.2017.2752720},
 }
 @article{legros2022instantaneous,
-  title={Instantaneous Frequency Estimation In Multi-Component Signals Using Stochastic {EM} Algorithm},
+  title={Instantaneous frequency and amplitude estimation in multicomponent signals using an em-based algorithm},
   author={Legros, Quentin and Fourer, Dominique and Meignen, Sylvain and Colominas, Marcelo A},
-  journal={arXiv preprint arXiv:2203.16334},
-  year={2022}
+  journal={IEEE Transactions on Signal Processing},
+  volume={72},
+  pages={1130--1140},
+  year={2024},
+  publisher={IEEE},
+  doi = {10.1109/TSP.2024.3361713},
 }
 
 @article{legros2022pb, title={Pseudo-Bayesian Approach for Robust Mode Detection and Extraction Based on the STFT}, volume={23}, ISSN={1424-8220}, url={http://dx.doi.org/10.3390/s23010085}, DOI={10.3390/s23010085}, number={1}, journal={Sensors}, publisher={MDPI AG}, author={Legros, Quentin and Fourer, Dominique}, year={2022}, month={Dec}, pages={85} }
@@ -76,7 +87,8 @@
   title={Time-Frequency Ridge Estimation of Multi-Component Signals using Sparse Modeling of Signal Innovation},
   author={Legros, Quentin and Fourer, Dominique},
   journal={arXiv preprint arXiv:2212.11343},
-  year={2022}
+  year={2022},
+  doi = {10.48550/arXiv.2212.11343},
 }
 
 @article{laurent2020novel,
@@ -86,7 +98,8 @@
   volume={27},
   pages={935--939},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/lsp.2020.2996905},
 }
 
 @article{colominas2020fully,
@@ -96,7 +109,8 @@
   volume={27},
   pages={620--624},
   year={2020},
[…]
