+++ paper.bib
@@ -166,7 +166,8 @@
   journal={Proceedings of Machine Learning and Systems},
   volume={2},
   pages={336--349},
-  year={2020}
+  year={2020},
+  doi = {10.48550/arXiv.1910.01500},
 }
 
 @article{hansen2021coco,
