+++ pyproject.toml
@@ -11,7 +11,7 @@ repository = "https://github.com/jmiramont/mcsm-benchs"
 include = ["LICENSE",]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.11"
+python = ">=3.10,<3.11"
 numpy = "^1.22.0"
 matplotlib = "^3.5.1"
 pandas = "^1.4.4"
@@ -25,6 +25,7 @@ plotly = "^5.10.0"
 parallelbar = "^2.4"
 pytest-cov = "^5.0.0"
 tqdm = "^4.66.4"
+
 [tool.poetry.group.docs]
 optional = true
 
@@ -36,8 +37,9 @@ sphinxcontrib-bibtex  = "^2.5.0"
 myst-parser =  "^2.0.0"
 sphinx-copybutton = "^0.5.2"
 insegel = "^1.3.1"
-sphinx-press-theme = "^0.8.0"
-sphinx-material = "^0.0.36"
+sphinxemoji = "^0.3.1"
+# sphinx-press-theme = "^0.8.0"
+# sphinx-material = "^0.0.36"
 
 # Dependencies for notebooks. "poetry install --with notebooks"
 [tool.poetry.group.notebooks]
