+++ docs/refs.bib
@@ -1,0 +1,955 @@
+@article{LaLa22Sub,
+	Author = {Ladgham, S. and Lachi\`eze-Rey, R.},
+	Date-Added = {2022-11-28 15:36:31 +0100},
+	Date-Modified = {2022-11-28 15:36:56 +0100},
+	Journal = {arXiv preprint arXiv:2209.04150},
+	Title = {Local repulsion of planar Gaussian critical points},
+	Year = {2022}}
+
+@book{HOUZ10,
+	Author = {Holden, H. and {\O}ksendal, B. and Ub{\o}e, J. and Zhang, T.},
+	Booktitle = {Stochastic partial differential equations},
+	Date-Added = {2022-11-25 14:00:20 +0100},
+	Date-Modified = {2022-11-25 14:00:20 +0100},
+	Edition = {Second},
+	Publisher = {Springer},
+	Title = {Stochastic partial differential equations},
+	Year = {2010}}
+
+@article{richards2013discrete,
+	Author = {Richards, Mark A},
+	Journal = {Georgia Institute of Technology, Tech. Rep},
+	Title = {The discretetime Fourier transform and discrete Fourier transform of windowed stationary white noise},
+	Year = {2013}}
+
+@article{meignen2021interference,
+	Author = {Meignen, Sylvain and Oberlin, Thomas},
+	Title = {On the Study of Interference and their Localization in the Time-Frequency Plane},
+	Year = {2021}}
+
+@article{donoho1994ideal,
+	Author = {Donoho, David L and Johnstone, Jain M},
+	Journal = {Biometrika},
+	Number = {3},
+	Pages = {425--455},
+	Publisher = {Oxford University Press},
+	Title = {Ideal spatial adaptation by wavelet shrinkage},
+	Volume = {81},
+	Year = {1994}}
+
+@article{donoho1995noising,
+  title={De-noising by soft-thresholding},
+  author={Donoho, David L},
+  journal={IEEE {T}ransactions on {I}nformation {T}heory},
+  volume={41},
+  number={3},
+  pages={613--627},
+  year={1995},
+  publisher={IEEE}
+}
+
+@article{meignen2021use,
+	Author = {Meignen, Sylvain and Pham, Duong-Hung and Colominas, Marcelo A},
+	Journal = {Signal Processing},
+	Pages = {107760},
+	Publisher = {Elsevier},
+	Title = {On the use of short-time fourier transform and synchrosqueezing-based demodulation for the retrieval of the modes of multicomponent signals},
+	Volume = {178},
+	Year = {2021}}
+
+@article{meignen2018retrieval,
+	Author = {Meignen, Sylvain and Pham, Duong-Hung},
+	Journal = {IEEE Transactions on Signal Processing},
+	Number = {23},
+	Pages = {6204--6215},
+	Publisher = {IEEE},
+	Title = {Retrieval of the modes of multicomponent signals from downsampled short-time fourier transform},
+	Volume = {66},
+	Year = {2018}}
+
+@book{grochenig2001foundations,
+	Author = {Gr{\"o}chenig, Karlheinz},
+	Publisher = {Springer Science \& Business Media},
+	Title = {Foundations of time-frequency analysis},
+	Year = {2001}}
+
+@article{escudero2021efficient,
+	Author = {Escudero, Luis Alberto and Feldheim, Naomi and Koliander, G{\"u}nther and Romero, Jos{\'e} Luis},
+	Journal = {arXiv preprint arXiv:2108.12921},
+	Title = {Efficient computation of the zeros of the Bargmann transform under additive white noise},
+	Year = {2021}}
+
+@article{flandrin2015time,
+	Author = {Flandrin, Patrick},
+	Journal = {IEEE Signal Processing Letters},
+	Number = {11},
[…]
