+++ docs/introduction.rst
@@ -1,0 +1,5 @@
+Introduction
+============
+
+.. include:: intro.md
+   :parser: myst_parser.sphinx_
