+++ docs/install.md
@@ -1,0 +1,5 @@
+# Installation using ```pip```
+
+```bash
+pip install mcsm-benchs
+```
