+++ pyproject.toml
@@ -38,6 +38,7 @@ myst-parser =  "^2.0.0"
 sphinx-copybutton = "^0.5.2"
 insegel = "^1.3.1"
 sphinxemoji = "^0.3.1"
+setuptools = "^75.8.0"
 # sphinx-press-theme = "^0.8.0"
 # sphinx-material = "^0.0.36"
 
