+++ _sources/intro.md.txt
@@ -1,4 +1,180 @@
-# Installation using ```pip```
-```bash
-pip install mcsm-benchs
+# What is `mcsm-benchs`?
+
+`mcsm-benchs` is a Python-based toolbox for benchmarking methods that aim at detecting, analyzing, and processing signals that are the addition of several individual components of interest, i.e. multi-component signals.
+
+This toolbox was born out of the necessity of simplifying systematic comparison of new approaches with existing ones in an area with vast amounts of methods like time-frequency representations of multi-component signals, which still remains a field of methodological innovation.
+Tools with similar goals are already being used in neighboring fields like optimization or machine learning.
+
+`mcsm-benchs` was designed to underpin collaborative benchmarks available through online repositories (see here a custom GitHub [template](https://github.com/jmiramont/collab-benchmark-template) to create your own).
+
+# Architecture and basic functionality
+`mcsm-benchs` was designed as a modular software, so that different modules can interact with each other. They may also be easily updated, or replaced, without affecting the other components in a benchmark.
+The object-oriented programming paradigm was extensively used, and the basic components of a benchmark, as well as other utilities provided by the library, are represented by different `Python` classes.
+
+
+```{figure} ./readme_figures/dessin_color_2.svg
+:alt: Block diagram describing `mcsm-benchs`classes and their interaction.
+:name: block-diagram
+:width: 700px
+:align: center
+Block diagram describing `mcsm-benchs`classes and their interaction.
+```
+
+Figure {numref}`block-diagram` shows a block diagram depicting the interactions between the main components of the toolbox and the inputs from the user.
+Optional input parameters are denoted with dotted lines.
+The three main classes of `mcsm-benchs` are: 1) the `SignalBank`class, 2) the `Benchmark` class and 3) the `ResultsInterpreter` class.
+Figure {numref}`block-diagram` also shows the four main attributes of the `Benchmark` class: 
+1. the noise generation function.
+2. the dictionary of methods to be compared.
+3. the dictionary of performance metrics
+4. the results.
+
+Following the pipeline in {numref}`block-diagram`, the user defines a set of signals to test, or generates them using the methods provided by the `SignalBank`class.
+Then, the simulation parameters, such as the length of the signals, the number of noise realizations to use, etc., are passed to the `Benchmark` class creator and a `Benchmark` object is instantiated.
+
+To run the benchmark, the class method [`run(...)`](https://jmiramont.github.io/mcsm-benchs/mcsm_benchs.html#mcsm_benchs.Benchmark.Benchmark.run) is called, which first executes the noise generation function to create a set of noise realizations.
+Then, these realizations are added to the signals provided by the user, yielding signal-plus-noise mixtures with the specified signal-to-noise ratios (SNR, in dB).
+The noisy signals are then fed to the methods provided by the user as function handlers in a `Python` dictionary of methods.
+Their outputs are in turn given to the performance metrics.
+The main final product of a benchmark created with `mcsm-benchs` is a two-dimensional data array, technically a Pandas `DataFrame`.
+Results can be processed by the `ResultsInterpreter` class, which generates output files and interactive figures to show in custom websites.
+
+**If a method raises an exception during execution time, the benchmarking process is not stopped.**
+Instead, the exception is caught internally, and a warning is shown in the standard output, indicating that the results of that iteration are replaced by `NaN`.
+Additionally, a log entry is saved with the method name and the noise realization that was used in that iteration.
+The log is a `Benchmark` attribute, and can then be helpful to reproduce any particular error during the benchmarking process.
+
+## Noise generation function
+
+A function handler to a noise generator method (or a dictionary of thse) can be passed by the user as an input parameter.
[…]
