+++ docs/mcsm_benchs.rst
@@ -1,53 +1,0 @@
-mcsm\_benchs package
-====================
-
-Submodules
-----------
-
-mcsm\_benchs.Benchmark module
------------------------------
-
-.. automodule:: mcsm_benchs.Benchmark
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-mcsm\_benchs.MatlabInterface module
------------------------------------
-
-.. automodule:: mcsm_benchs.MatlabInterface
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-mcsm\_benchs.ResultsInterpreter module
---------------------------------------
-
-.. automodule:: mcsm_benchs.ResultsInterpreter
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-mcsm\_benchs.SignalBank module
-------------------------------
-
-.. automodule:: mcsm_benchs.SignalBank
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-mcsm\_benchs.benchmark\_utils module
-------------------------------------
-
-.. automodule:: mcsm_benchs.benchmark_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: mcsm_benchs
-   :members:
-   :undoc-members:
-   :show-inheritance:
