+++ .github/workflows/build_docs.yml
@@ -46,6 +46,7 @@ jobs:
       - name: Build docs
         run: |
           cd docs
+          poetry run sphinx-apidoc -o . ../mcsm_benchs 
           poetry run make clean
           poetry run make html     
         
