+++ docs/collab_benchs.rst
@@ -1,0 +1,5 @@
+Introduction
+============
+
+.. include:: collab.md
+   :parser: myst_parser.sphinx_
