+++ docs/intro.md
@@ -106,10 +106,14 @@ Spectrograms of some of the signals generated by the `SignalBank`class.
 
 Noisy multi-component synthetic signals enable quantitative evaluation of the performance of a method, since both the noiseless version of the signal and the noise are known.
 The `SignalBank`class included in `mcsm-benchs` can be used to generate more than $20$ multi-component signals with a variety of TF structures, which serves the purpose of standardizing benchmarks across different users by providing a common input to the methods.
-{numref}`spectrograms` displays the spectrograms of some of the available signals produced by the `SignalBank`class.
+{numref}`spectrograms` displays the spectrograms of some of the available signals produced by the `SignalBank` class.
 
 More specifically, a discrete multi-component, amplitude-modulated, frequency-modulated (AM-FM) signal $s$ can be written as
- $$  s[n] = \sum_{j=1}^{J} s_{j}[n], \text{ with } s_{j}[n] = a_{j}[n]e^{i \phi_{j}[n]}$$
+
+```{math}
+s[n] = \sum_{j=1}^{J} s_{j}[n], \text{ with } s_{j}[n] = a_{j}[n]e^{i \phi_{j}[n]}
+```
+
 where $n=0,1,\dots,N-1$ is the discrete time index, $J$ is the number of signal components, $s_{j}$ is an AM-FM component, and $a_{j}$ and $\phi_{j}$ are the instantaneous amplitude and phase functions of the $j$-th mode, respectively.
 Each mode is associated with an instantaneous frequency given by $\frac{1}{2\pi}\varphi^{\prime}(nT_{s})$, $n=0,1,\dots,N-1$, where $\varphi$ is the continuous-time counterpart of the phase function corresponding to the $j$-$th$ mode and $T_{s}$ is the sampling period.
 
@@ -127,7 +131,7 @@ In practice, `Signal` objects behave like regular `numpy` arrays, but they also 
 - `comps`: `numpy` array with the individual components (as independent signals).
 - `instf`: `numpy` array with the instantaneous frequency of each component.
 
-Figure {numref}`signal-attributes` illustrates the attributes of a `Signal` class instance: the signal, its individual components as well as their instantaneous frequencies, and the number of components present in the signal for each instant.
+{numref}`signal-attributes` illustrates the attributes of a `Signal` class instance: the signal, its individual components as well as their instantaneous frequencies, and the number of components present in the signal for each instant.
 
 ```{figure} ./readme_figures/signals_ifs.svg
 :alt: Attributes of the `Signal`class.
