+++ pyproject.toml
@@ -20,9 +20,7 @@ seaborn = ">=0.12.0"
 pytest = "^7.0.1"
 tabulate = "^0.8.9"
 plotly = "^5.10.0"
-
-# Dependencies for documentation. "poetry install --with docs"
-parallelbar = "^2.4"
+# parallelbar = "^2.4"
 pytest-cov = "^5.0.0"
 tqdm = "^4.66.4"
 
