+++ pyproject.toml
@@ -24,6 +24,7 @@ plotly = "^5.10.0"
 pytest-cov = "^5.0.0"
 tqdm = "^4.66.4"
 
+
 [tool.poetry.group.docs]
 optional = true
 
@@ -37,6 +38,7 @@ sphinx-copybutton = "^0.5.2"
 insegel = "^1.3.1"
 sphinxemoji = "^0.3.1"
 setuptools = "^75.8.0"
+sphinx-math-dollar = "^1.2.1"
 # sphinx-press-theme = "^0.8.0"
 # sphinx-material = "^0.0.36"
 
