+++ paper.md
@@ -41,22 +41,29 @@ MCS processing is an area with a long history which, however, remains a field of
 In such a context, systematically comparing existing methods while keeping a record of how they performed on a predefined set of representative problems, i.e. benchmarking, can shed light on novel avenues of research and set clear baselines for new approaches.
 This is a widely adopted strategy in neighboring fields such as optimization [@bartz2020benchmarking; @hansen2021coco; @benchopt] and machine learning [@mattson2020mlperf], yet there are no benchmarking tools for MCS processing.
 `mcsm-benchs` is an open-source Python library for creating reproducible benchmarks of methods that aim at extracting information from MCSs.
-`Benchmark` objects can be created by simply passing a series of simulation parameters, a dictionary of methods and a set of performance metrics, while a `SignalBank` can be used to standardize comparisons across benchmarks.
-This latter can synthesize several MCSs as objects from a custom `Signal` class that behave like regular arrays in Python but also contain information about the components.
+`Benchmark` objects can be created by simply passing a series of simulation parameters, a dictionary of methods and a set of performance metrics, while a `SignalBank` instance can be used to standardize comparisons across benchmarks.
+The `SignalBank` class can synthesize several MCSs as objects from a custom `Signal` class that behave like regular arrays in Python but also contain information about the components.
 
-Additionally, `mcsm-benchs` includes a `ResultsInterpreter` class that can produce human-readable reports, the functionality of which underpins *collaborative* benchmarks.
+Additionally, `mcsm-benchs` includes a `ResultsInterpreter` class that can produce human-readable reports, the functionality of which underpins *collaborative* benchmarks [@benchopt].
 These are based on online repositories (see [here](https://jmiramont.github.io/benchmarks-detection-denoising/results_denoising.html) for an example) and can be periodically updated by members of the research community, fostering open and collaborative science.
-Several examples are given in the documentation, as well as a GitHub [repository template](https://github.com/jmiramont/collab-benchmark-template) that relies heavily on `mcsm-benchs` and continuous integration/deployment workflows, in order to automatize the process of publishing new benchmarks.
+Several examples are given in the [documentation](https://jmiramont.github.io/mcsm-benchs/), as well as a GitHub [repository template](https://github.com/jmiramont/collab-benchmark-template) that relies heavily on `mcsm-benchs` and continuous integration/deployment workflows, in order to automatize the process of publishing new benchmarks.
 
 # Statement of need
 
 `mcsm-benchs` brings to the table a common framework to easily carry out extensive comparisons between MCS-based approaches in a unified and objective way.
-For instance, `mcsm-benchs` was used to compare statistical methods for signal detection [@miramont2022public] and signal denoising of synthetic and real signals under different scenarios, such as white noise and real-world noises [@miramont2024].
-Additionally, benchmarks of instantaneous frequency and component estimation methods where introduced in @miramont2023eusipco.
+It can be used to benchmark any number of approaches and create clear baselines for new methods that are accessible to the whole research community.
 
-While `mcsm-benchs` was mainly designed with MCSs in mind, methods are always treated like black boxes, i.e. only the shape of the input and output arguments are constrained by the toolbox.
-Therefore, `mcsm-benchs` can potentially be used to compare any signal processing algorithm as long as its output matches the inputs of a given performance metric.
-Finally, to ease the adoption of our tool by the community, `mcsm-benchs` also supports methods coded in `Octave`/`Matlab`, so that these can be seamlessly integrated into new Python-based benchmarks.
+The toolbox is versatile enough to allow comparisons between many kinds of methods.
[…]
