+++ pyproject.toml
@@ -11,8 +11,8 @@ repository = "https://github.com/jmiramont/mcsm-benchs"
 include = ["LICENSE",]
 
 [tool.poetry.dependencies]
-python = ">=3.9"
-numpy = "^1.22.0"
+python = ">=3.10"
+numpy = "^2.1.0"
 matplotlib = "^3.5.1"
 # pandas = "1.4.4"
 scipy = "^1.7.3"
