+++ pyproject.toml
@@ -18,11 +18,12 @@ matplotlib = "^3.5.1"
 scipy = "^1.7.3"
 seaborn = ">=0.12.0"
 pytest = "^7.0.1"
-tabulate = "^0.8.9"
+# tabulate = "^0.8.9"
 plotly = "^5.10.0"
 # parallelbar = "^2.4"
 pytest-cov = "^5.0.0"
 tqdm = "^4.66.4"
+pandas = "^2.2.3"
 
 
 [tool.poetry.group.docs]
