### 2025-05-07 #8175 @mstimberg
at the top of a new comment in this thread.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/8175#issue-3045196259).
