### 2025-06-12 jmiramont/mcsm-benchs#2 @tsbinns
[JOSS] Review Suggestions
I am one of the reviewers for your submission to JOSS (https://github.com/openjournals/joss-reviews/issues/8175).
I was very happy with the paper and software, which has a clear benefit for the signal analysis community. There are only a few small points I suggest are addressed:

### Major

- [ ] For installing the package for development, I suggest adding a step in the instructions that the repository should first be cloned and navigated to before trying to install via poetry. I expect most users wishing to install the package in development mode will realise this a pre-requisite, nonetheless it would be useful to explicitly state that the repository must first be cloned.

### Minor

- [ ] I would find it useful to state in the documentation that Python 3.10 is required. Since this is ~3 years old, new - and many existing environments - will be using newer versions of Python that are not compatible. Stating the version requireement in the documentation can help users to better plan how they will access the package before trying to install anything.

### Major

- [ ] The guidelines for adding new benchmarks are put together very well. It would be good to also include guidelines for how users can report issues and contribute to the package (such as by linking to the repository's issues and pull request pages).
- [ ] A code of conduct should also be in place that sets out the expected behaviour for those involved and wishing to engage in the project. JOSS has a template [code of conduct](https://github.com/openjournals/joss/blob/main/CODE_OF_CONDUCT.md) that can be used.
- [ ] It would be useful to add instructions for how users can run the unit tests manually ([example](https://mne.tools/dev/development/contributing.html#running-the-test-suite)).

### Minor

- [ ] For package contributions, it would be useful to mention that new features should be accompanied by unit tests and documentation. Instructions for how to run the unit tests (as above) and build the documentation ([example](https://mne.tools/dev/development/contributing.html#building-the-documentation)) would also be beneficial.

### Major

- [ ] In the section "Generating plots with the Results Interpreter" of [`demo_benchmark`](https://github.com/jmiramont/mcsm-benchs/blob/main/notebooks/demo_benchmark.ipynb), the plots are not rendered in the online documentation (but admittedly can be viewed in the notebooks).
- [ ] In the section "Generating plots with the Results Interpreter" of [`demo_other_user_inputs`](https://github.com/jmiramont/mcsm-benchs/blob/main/notebooks/demo_other_user_inputs.ipynb), the plots are not rendered in the online documentation (but admittedly can be viewed in the notebooks).

### Minor

- [ ] In the section "Creating a dictionary of methods" of [`demo_other_user_inputs`](https://github.com/jmiramont/mcsm-benchs/blob/main/notebooks/demo_other_user_inputs.ipynb), the definitions of the thresholding functions are missing.

## Other documentation

### Minor

- [ ] There are a number of small spelling mistaked in the documentation. I will submit a PR with these fixed that you can use directly or adapt as you see fit.
- [ ] There is a small irregularity in how figures are referenced in [`docs/intro`](https://github.com/jmiramont/mcsm-benchs/blob/main/docs/intro.md) (e.g., "Figure Fig. 1"). I will submit a PR with these fixed that you can use directly or adapt as you see fit.
- [ ] In [`docs/intro`](https://github.com/jmiramont/mcsm-benchs/blob/main/docs/intro.md) line 160, their is a reference to a "results" section that does not render properly.
- [ ] In [`docs/intro`](https://github.com/jmiramont/mcsm-benchs/blob/main/docs/intro.md) line 162, their is a `\code` tag that does not render properly.
- [ ] In [`docs/collab`](https://github.com/jmiramont/mcsm-benchs/blob/main/docs/collab.md) line 4, their is an `\emph` tag that does not render properly.

Altogether the package is well documented and provides a set of resources with a very clear benefit for benchmarking signal analysis methods. Please let me know if you have any questions about the above points.
