### 2025-06-16 f750b94
Updating installation instructions.
