### 2025-06-16 f9ffaad
Adding contribution guidelines.
