+++ .github/workflows/build_docs.yml
@@ -61,18 +61,3 @@ jobs:
           BRANCH: gh-pages
           FOLDER: docs/_build/html
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-      # - name: Publish results
-      #   env:
-      #     SRC_FOLDER_PATH: 'results'
-      #     TARGET_BRANCH: 'gh-pages'
-      #   run: |
-      #     files=$(find $SRC_FOLDER_PATH -type f) # get the file list
-      #     git config --global user.name 'GitHub Action'
-      #     git config --global user.email 'action@github.com'
-      #     git fetch                         # fetch branches
-      #     git checkout $TARGET_BRANCH       # checkout to your branch
-      #     git checkout ${GITHUB_REF##*/} -- $files # copy files from the source branch
-      #     git add -A
-      #     git diff-index --quiet HEAD ||  git commit -am "deploy files"  # commit to the repository (ignore if no modification)
-      #     git push origin $TARGET_BRANCH # push to remote branch
