+++ docs/contrib.rst
@@ -1,0 +1,5 @@
+Installation
+============
+
+.. include:: contributions.md
+   :parser: myst_parser.sphinx_
