### 2025-06-16 ee2af2d
Adding contribution guidelines
