### 2025-06-23 6e784d3
Fixing the plots in documentation.
