+++ pyproject.toml
@@ -25,7 +25,6 @@ pytest-cov = "^5.0.0"
 tqdm = "^4.66.4"
 pandas = "^2.2.3"
 
-
 [tool.poetry.group.docs]
 optional = true
 
@@ -51,6 +50,7 @@ optional = true
 ipykernel = {version = "^6.6.1"}
 nbformat = "^5.9.2"
 pesq = "^0.0.4"
+tabulate = "^0.9.0"
 
 # # Install Matlab-based tools with "poetry install --with matlab_tools"
 # [tool.poetry.group.matlab_tools]
