### 2025-07-15 jmiramont/mcsm-benchs#5 @nmy2103
[JOSS] Reviewer Feedback
Dear @jmiramont,

I am one of the reviewers for your submission to JOSS (https://github.com/openjournals/joss-reviews/issues/8175). I apologize for my delay in providing feedback to you.

Overall, I was pleased with `mcsm-benchs` and see its utility to researchers in signal processing and commitment to transparency and reproducibility. There are, however, some issues that I hope you could address.

### Software Paper

- [x] **Summary/Statement of Need:** While they are implied in your submission, your claims for the utility of your software can be more directly and clearly described in the "Statement of Need" section. I believe you alluded to the research gap in your "Summary" section:

  > In such a context, systematically comparing existing methods while keeping a record of how they performed on a predefined set of representative problems, i.e. benchmarking, can shed light on novel avenues of research and set clear baselines for new approaches. This is a widely adopted strategy in neighboring fields such as optimization and machine learning, yet there are no benchmarking tools for MCS processing.

  I recommend **restructuring or reorganizing the ideas** presented in these sections for more clarity.

- [x] **State of the Field:** The paper does not explicitly state what existing tools or packages (if any) researchers currently use to benchmark MCS processing methods. Providing **an example of a typical pipeline or describing common ad hoc approaches** would help situate `mcsm-benchs` within the existing research ecosystem and better highlight its contribution.

### Functionality

- [x] **Example Usage:** In both notebooks `demo_benchmark.ipynb` and `demo_other_user_inputs.ipynb`, only the `task = 'denoising'` workflow is demonstrated. While this provides a good baseline, the other supported tasks (`'detection'`, `'component_denoising'`, and `'inst_frequency'`) are not shown in practice.

  For example, when attempting to use `task = 'inst_frequency'` with a Hilbert transform method on a `'LinearChirp'` signal in `demo_benchmark.ipynb`, I encountered `NaN` values in the results. While your notebook does explain the expected output format for 'denoising' and 'detection', it doesn't specify what shape/structure is required for instantaneous frequency estimation, especially in the case of multicomponent signals. As a result, it may be easy for a user to implement a method that appears to work but fails silently during evaluation. I highly recommend including minimal **working examples for each `task` mode other than `'denoising'`**, which would help users understand how to configure their methods correctly and interpret the benchmarking results.
