### 2025-07-26 724bece
Adding new notebooks with examples.
