### 2025-07-26 d6f6ef7
Updating notebooks
