+++ _sources/index.rst.txt
@@ -34,6 +34,7 @@ The goal of this toolbox is providing the signal-processing community with a com
    notebooks/demo_noise_function.ipynb
    notebooks/demo_quering_signals.ipynb
    notebooks/demo_if_estimation.ipynb
+   notebooks/demo_component_estim.ipynb 
    notebooks/demo_other_user_inputs.ipynb
 
 .. toctree::
