### 2025-09-14 #8175 @nmy2103
@mstimberg @danielskatz Hello, Marcel and Daniel. I hope you are well and apologize for my delay. Yes, all comments have been sufficiently addressed by @jmiramont. I believe their software and paper meet the requirements for publication in JOSS.
