### 2025-09-15 #8175 @mstimberg
Many thanks for your confirmation @nmy2103, and thanks again to @tsbinns for their review. With both reviewers recommending acceptance, we can move to the final editorial checks. @jmiramont please take care of the "Authors tasks" in the checklist below, and comment back here when they are done. In parallel, I will verify that the manuscript matches our editorial policies.
