### 2025-09-15 8e73502
Create CITATION.cff
