+++ CITATION.cff
@@ -1,0 +1,21 @@
+cff-version: 0.1.3
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: "Juan M."
+  given-names: "Miramont"
+  orcid: "https://orcid.org/0000-0002-3847-7811"
+- family-names: "Remi"
+  given-names: "Bardenet"
+  orcid: "https://orcid.org/0000-0002-1094-9493"
+- family-names: "Pierre"
+  given-names: "Chainais"
+  orcid: "https://orcid.org/0000-0003-4377-7584"
+- family-names: "François"
+  given-names: "Auger"
+  orcid: "https://orcid.org/0000-0001-9158-1784"
+
+title: "mcsm-benchs: A Toolbox for Benchmarking Multi-Component Signal Analysis Methods"
+version: 0.1.3
+doi: 10.5281/zenodo.17122672
+date-released: 2025-09-15
+url: "https://github.com/jmiramont/mcsm-benchs"
