### 2025-09-17 2ffbf32
Minor changes for JOSS publication
