+++ paper.bib
@@ -17,14 +17,13 @@
 }
 
 @inproceedings{kreme2020filtering,
-  title={Filtering out time-frequency areas using Gabor multipliers},
+  title={Filtering out time-frequency areas using {Gabor} multipliers},
   author={Kr{\'e}m{\'e}, A Marina and Emiya, Valentin and Chaux, Caroline and Torresani, Bruno},
   booktitle={ICASSP 2020-2020 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
   pages={5790--5794},
   year={2020},
   organization={IEEE},
   doi = {10.1109/icassp40776.2020.9053482},
-
 }
 
 @article{pascal2022covariant,
@@ -71,7 +70,7 @@
   doi = {10.1109/tsp.2017.2752720},
 }
 @article{legros2022instantaneous,
-  title={Instantaneous frequency and amplitude estimation in multicomponent signals using an em-based algorithm},
+  title={Instantaneous frequency and amplitude estimation in multicomponent signals using an EM-based algorithm},
   author={Legros, Quentin and Fourer, Dominique and Meignen, Sylvain and Colominas, Marcelo A},
   journal={IEEE Transactions on Signal Processing},
   volume={72},
@@ -86,8 +85,11 @@
 @article{legros2022time,
   title={Time-Frequency Ridge Estimation of Multi-Component Signals using Sparse Modeling of Signal Innovation},
   author={Legros, Quentin and Fourer, Dominique},
-  journal={arXiv preprint arXiv:2212.11343},
   year={2022},
+  eprint={2212.11343},
+  archivePrefix={arXiv},
+  publisher = {arXiv}, 
+  primaryClass={math.NA},
   doi = {10.48550/arXiv.2212.11343},
 }
 
@@ -152,21 +154,27 @@
   doi = {10.23919/eusipco58844.2023.10290093},
 }
 
-@article{bartz2020benchmarking,
+@misc{bartz2020benchmarking,
   title={Benchmarking in optimization: Best practice and open issues},
-  author={Bartz-Beielstein, Thomas and Doerr, Carola and Berg, Daan van den and Bossek, Jakob and Chandrasekaran, Sowmya and Eftimov, Tome and Fischbach, Andreas and Kerschke, Pascal and La Cava, William and Lopez-Ibanez, Manuel and others},
-  journal={arXiv preprint arXiv:2007.03488},
+  author={Bartz-Beielstein, Thomas and Doerr, Carola and Berg, Daan van den and Bossek, Jakob and Chandrasekaran, Sowmya and Eftimov, Tome and Fischbach, Andreas and Kerschke, Pascal and La Cava, William and Lopez-Ibanez, Manuel and others},  
   year={2020},
+  eprint={2007.03488},
+  archivePrefix={arXiv},
+  primaryClass={cs.NE},
+  publisher = {arXiv},
   doi={10.48550/arXiv.2007.03488},
 }
 
-@inproceedings{mattson2020mlperf,
+@misc{mattson2020mlperf,
   title={{MLPerf} training benchmark},
   author={Mattson, Peter and Cheng, Christine and Diamos, Gregory and Coleman, Cody and Micikevicius, Paulius and Patterson, David and Tang, Hanlin and Wei, Gu-Yeon and Bailis, Peter and Bittorf, Victor and others},
   journal={Proceedings of Machine Learning and Systems},
   volume={2},
   pages={336--349},
   year={2020},
+  eprint={1910.01500},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG},
   doi = {10.48550/arXiv.1910.01500},
 }
 
@@ -181,13 +189,14 @@ volume = {36},
 year = 2021
 }
 
[…]
