### 2025-09-19 #8175 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
