### 2025-09-22 #8175 @danielskatz
Congratulations to @jmiramont (J. M. Miramont) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers, and we couldn't do this without you
